Move Drawer to lab
