--- packages/cli/templates/blocks/components/Drawer/DrawerBottomSheetDrawer.doc.mjs
@@ -1,14 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'Drawer',
-  name: 'Drawer — Bottom Sheet',
-  displayName: 'Drawer — Bottom Sheet',
-  description:
-    'Bottom sheet pattern: side="bottom" slides a full-width sheet from the bottom edge with a height budget (size accepts any CSS length).',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['Button', 'CheckboxInput', 'Drawer', 'Heading', 'Section', 'Stack', 'Text'],
-};
