--- packages/cli/templates/blocks/components/Drawer/DrawerRowInspectorDrawer.doc.mjs
@@ -1,14 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'Drawer',
-  name: 'Drawer — Row Inspector',
-  displayName: 'Drawer — Row Inspector',
-  description:
-    'Master-detail pattern: clicking a row opens a non-modal inspector drawer while the list stays interactive.',
-  isReady: true,
-  aspectRatio: 4 / 3,
-  componentsUsed: ['Button', 'Divider', 'Drawer', 'Heading', 'Section', 'Stack', 'Text'],
-};
