--- packages/cli/templates/blocks/components/Drawer/DrawerShowcase.doc.mjs
@@ -1,15 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'Drawer',
-  name: 'Drawer',
-  displayName: 'Drawer',
-  description:
-    'A button that opens a right-side inspector drawer with detail content.',
-  isReady: true,
-  isShowcase: true,
-  aspectRatio: 1,
-  componentsUsed: ['Button', 'Divider', 'Drawer', 'Heading', 'Section', 'Stack', 'Text'],
-};
