+++ packages/core/src/index.ts
@@ -87,7 +87,6 @@ export * from './TreeList';
 export * from './Kbd';
 export * from './AlertDialog';
 export * from './Dialog';
-export * from './Drawer';
 export * from './ContextMenu';
 export * from './DropdownMenu';
 export * from './MoreMenu';
