+++ packages/lab/src/Drawer/Drawer.doc.mjs
@@ -1,5 +1,5 @@
 // Copyright (c) Meta Platforms, Inc. and affiliates.
-/** @type {import('../docs-types').ComponentDoc} */
+/** @type {import('../../core/src/docs-types').ComponentDoc} */
 
 export const docs = {
   name: 'Drawer',
