+++ packages/lab/src/index.ts
@@ -28,6 +28,9 @@ export {
   type ChatReasoningProps,
 } from './ChatReasoning/ChatReasoning';
 
+// Drawer — experimental overlay panel
+export {Drawer, type DrawerProps} from './Drawer';
+
 // Schedule — experimental full-calendar views
 export * from './Schedule';
